+++ lib/disposable/twin/parent.rb
@@ -3,6 +3,7 @@ module Disposable::Twin::Parent
     includer.property(:parent, virtual: true)
   end
 
+  # FIXME: for collections, this will merge options for every element.
   def build_twin(dfn, value, options={})
     super(dfn, value, options.merge(parent: self))
   end