+++ test/twin/collection_test.rb
@@ -218,7 +218,8 @@ class CollectionUnitTest < MiniTest::Spec
     Album = Struct.new(:id, :name, :songs, :artist)
   end
 
-  let(:collection) { Disposable::Twin::Collection.new(Disposable::Twin::Twinner.new(Twin::Song.definitions.get(:album)), []) }
+  # THIS is why private tests suck!
+  let(:collection) { Disposable::Twin::Collection.new(Disposable::Twin::Twinner.new(Twin::Song.new(OpenStruct.new), Twin::Song.definitions.get(:album)), []) }
 
   # #insert(index, model)
   it do